## Changes

There was a flaw with how `rehype-optimize-static.ts` worked before. Take this page for example https://docs.astro.build/en/concepts/why-astro/, which is a static page with no components.

Before, the rehype plugin will only transform to something like this:
```jsx
<_components.h1 set:html="Why Astro?" /> 
<_components.p set:html="<strong>Astro</strong> is the web framework..." />
<_components.h1 set:html="Features" /> 
// ...
```

With this PR, it can go one step further and transform as:

```jsx
<Fragment set:html="<h1>Why Astro?</h1>\n<strong>Astro..." />
```

Which is only a single line and a single string, which simplifies the AST nodes and runtime rendering.

Performance:

<table>
<tr>
 <td>
 <td>Rollup build
 <td>Page rendering
<tr>
 <td>Before
 <td>86s
 <td>46s
<tr>
 <td>After
 <td>83s
 <td>34s
<tr>
 <td>Comparison
 <td><b>-3s (3.5% faster)</b>
 <td><b>-12s (26% faster)</b>
</table>

A bit dissapointed of the Rollup build time change, but I'll take the rendering improvement 😄 

## Testing

The `mdx-optimize` playground that should still pass. And I tested in the Astro docs repo.

## Docs

Added a changeset and updated the internal README explaining the new code.