@bluwy 2024-04-30
Thanks, I'll add a test 👍 There's currently the `mdx-optimize.test.js` file but it's indeed a bit simple.