@bluwy 2024-04-30
I added a unit test. It now reveals that it generates some cursed JSX code that has attribute values with quotes extending multiple lines, but it still seems to parse properly and we _don't have_ to manually escape the new lines, so leaving it like so for now.